extracted logout button to a separate component Co-authored-by: Lena Krasnova <28540894+olkras03@users.noreply.github.com>
