+++ client/src/components/auth/Logout-button.js
@@ -1,0 +1,33 @@
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import { connect } from "react-redux";
+import { logoutUser } from "../../actions/authActions";
+
+
+class Logout extends Component {
+  onLogoutClick = e => {
+    e.preventDefault();
+    this.props.logoutUser();
+  };
+
+  render() {
+    return (
+      <div>
+        <button onClick={this.onLogoutClick} className="logout">
+          Logout
+        </button>
+      </div>
+    );
+  }
+}
+
+Logout.propTypes = {
+    logoutUser: PropTypes.func.isRequired,
+    auth: PropTypes.object.isRequired
+  };
+  
+  const mapStateToProps = state => ({
+    auth: state.auth
+  });
+  
+  export default connect(mapStateToProps, { logoutUser })(Logout);
