+++ client/src/components/dashboards/Dashboard.js
@@ -5,13 +5,9 @@ import { logoutUser } from "../../actions/authActions";
 import Holiday from "./Holiday";
 import Holidays from "./Holidays";
 import './dashboard.scss'
+import Logout from "../auth/Logout-button"
 
 class Dashboard extends Component {
-
-  onLogoutClick = e => {
-    e.preventDefault();
-    this.props.logoutUser();
-  };
 
   onChange = event => {
     this.setState({ [event.target.name]: event.target.value });
@@ -26,11 +22,7 @@ class Dashboard extends Component {
         </div>
         <div className="row">
           <div className="Dashboard">
-            <div className="row">
-              <button onClick={this.onLogoutClick} className="logout">
-                Logout
-              </button>
-            </div>
+            <Logout />
           </div>
         </div>
       </div>
@@ -38,13 +30,14 @@ class Dashboard extends Component {
   }
 }
 
-Dashboard.propTypes = {
-  logoutUser: PropTypes.func.isRequired,
-  auth: PropTypes.object.isRequired
-};
+// Dashboard.propTypes = {
+//   logoutUser: PropTypes.func.isRequired,
+//   auth: PropTypes.object.isRequired
+// };
 
-const mapStateToProps = state => ({
-  auth: state.auth
-});
+// const mapStateToProps = state => ({
+//   auth: state.auth
+// });
 
-export default connect(mapStateToProps, { logoutUser })(Dashboard);
+export default Dashboard;
+// export default connect(mapStateToProps, { logoutUser })(Dashboard);
